requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.0
aiohttp==3.9.0